download via api end point to api module
