+++ ytd_web/get_mulimedia.py
@@ -26,14 +26,3 @@ def file_get_aud():
             return render_template("went_wrong.html", exception=e, links=StaticLinks)
     else:
         return redirect("/home")
-
-@get_miltimedia_page.route("/download/path", methods=['GET'])
-def download_from_link():
-    protected_files = ["secrets.key"]
-    for protected_file in protected_files:
-        if protected_file in request.args.get('file_name'):
-            return {'status': 403, 'description': "Illegal file name!"}, 403
-    try:
-        return send_file(request.args.get('file_name'), as_attachment=True)
-    except FileNotFoundError:
-        return {'status': 404, 'description': "link expired!"}, 404
